feat: Refactor App component and add new sections for Contact, Footer, Services, and Clients
