+++ src/App.tsx
@@ -1,73 +1,21 @@
 import './App.css';
-import logo from './assets/logo-removebg-preview.png';
-import { useState, useEffect, useRef } from 'react';
+import { useState } from 'react';
 import NavMenu from './components/NavMenu';
 import Testimonios from './components/Testimonios';
-const testimoniosIniciales = [
-  {
-    nombre: 'Juan Pérez',
-    texto: 'El equipo de NOESIS superó nuestras expectativas. Su atención al detalle y profesionalismo fue excepcional.'
-  },
-  {
-    nombre: 'María Gómez',
-    texto: 'Gracias a NOESIS, nuestra empresa ahora cuenta con una plataforma digital robusta y moderna.'
-  },
-  {
-    nombre: 'Carlos Ruiz',
-    texto: 'Excelente servicio y acompañamiento en todo el proceso de desarrollo. ¡Muy recomendados!'
-  },
-  {
-    nombre: 'Ana Torres',
-    texto: 'NOESIS nos ayudó a digitalizar procesos clave y mejorar la eficiencia de nuestro equipo.'
-  },
-  {
-    nombre: 'Luis Fernández',
-    texto: 'Gran calidad humana y técnica. Siempre atentos a nuestras necesidades.'
-  }
-];
-const proyectosIniciales = [
-  {
-    nombre: 'Plataforma de gestión educativa',
-    descripcion: 'Sistema web para administración de colegios y universidades.',
-    imagen: 'https://albert-z00z.github.io/Portfolio1/images/NumSecret.png',
-    tecnologias: ['React', 'Node.js', 'MongoDB'],
-    url: 'https://albert-z00z.github.io/juego-secreto/'
-  },
-  {
-    nombre: 'App de reservas para restaurantes',
-    descripcion: 'Aplicación móvil para gestionar reservas y pedidos en restaurantes.',
-    imagen: 'https://via.placeholder.com/320x180/cccccc/888888?text=Proyecto+2',
-    tecnologias: ['React Native', 'Firebase'],
-    url: 'https://ejemplo2.com'
-  },
-  {
-    nombre: 'Integración de sistemas ERP',
-    descripcion: 'Solución personalizada para integrar procesos empresariales.',
-    imagen: 'https://via.placeholder.com/320x180/cccccc/888888?text=Proyecto+3',
-    tecnologias: ['Python', 'Django', 'PostgreSQL'],
-    url: 'https://ejemplo3.com'
-  },
-  // Para agregar más proyectos, seguir el formato de los anteriores
-];
+import Contacto from './components/Contacto';
+import Footer from './components/Footer';
+import { proyectosIniciales } from './data/proyectos';
+import { testimoniosIniciales } from './data/testimonios';
+import Servicios from './components/Servicios';
+import Clientes from './components/Clientes';
 
-const logosGenericos = [
-  'https://upload.wikimedia.org/wikipedia/commons/6/69/LEGO_logo.svg',
-  'https://upload.wikimedia.org/wikipedia/commons/6/6a/JavaScript-logo.png',
-  'https://upload.wikimedia.org/wikipedia/commons/2/2f/Google_2015_logo.svg',
-  'https://upload.wikimedia.org/wikipedia/commons/5/51/Facebook_f_logo_%282019%29.svg',
-  'https://upload.wikimedia.org/wikipedia/commons/0/08/Netflix_2015_logo.svg'
-];
 
-// Slider de testimonios tipo carrusel 3 tarjetas
+
+
 
 
 function App() {
-  // Proyectos dinámicos
   const [proyectos] = useState(proyectosIniciales);
-
-  // Marquee infinito: duplicar logos para que siempre haya flujo
-  const marqueeLogos = [...logosGenericos, ...logosGenericos, ...logosGenericos, ...logosGenericos, ...logosGenericos];
-
   return (
     <div className="noesis-landing">
       <NavMenu/>
@@ -88,17 +36,7 @@ function App() {
       </section>
       {/* Logos de clientes - Marquee infinito */}
       <section className="clientes" id="clientes">
-        <div className="logos-titulo"><em>La gente está hablando</em></div>
-        <div className="marquee">
-          <div className="marquee-inner slow-marquee">
-            {marqueeLogos.map((src, i) => (
-              <img src={src} alt={`Logo cliente ${i+1}`} key={i} className="logo-cliente" />
-            ))}
-          </div>
-            <svg className="hero-wave" width="100%" height="80" viewBox="0 0 1440 80" fill="none" xmlns="http://www.w3.org/2000/svg">
-          <path d="M0,40 Q360,0 720,40 Q1080,80 1440,40" stroke="#bbb" strokeWidth="2" fill="none"/>
-        </svg>
-        </div>
+        <Clientes/>
       </section>
       {/* Testimonios - Carrusel 3 tarjetas */}
       <section className="testimonios">
@@ -110,18 +48,7 @@ function App() {
       </section>
       {/* Servicios */}
       <section className="servicios" id="servicios">
-        
-        <h3>Nuestros servicios</h3>
-        
-        <ul className="servicios-lista">
-          <li>Desarrollo web a medida</li>
-          <li>Aplicaciones móviles</li>
-          <li>Consultoría tecnológica</li>
-          <li>Integraciones y automatización</li>
-        </ul>
-        <svg className="servicios-wave" width="100%" height="60" viewBox="0 0 1440 60" fill="none" xmlns="http://www.w3.org/2000/svg">
-          <path d="M0,30 Q360,60 720,30 Q1080,0 1440,30" stroke="#bbb" strokeWidth="2" fill="none"/>
-        </svg>
+        <Servicios/>
       </section>
       {/* Proyectos en filas de 3, con imagen, tecnologías y botón */}
       <section className="proyectos" id="proyectos">
@@ -157,35 +84,8 @@ function App() {
         </div>
       </section>
       {/* Contacto */}
-      <section className="contacto" id="contacto">
-        <h3>¿Tienes preguntas? ¡Contáctanos!</h3>
-        <form className="form-contacto" onSubmit={e => e.preventDefault()}>
-          <input type="text" placeholder="Nombre" required />
-          <input type="email" placeholder="Correo electrónico" required />
-          <textarea placeholder="¿En qué podemos ayudarte?" required />
-          <button type="submit">Enviar</button>
-        </form>
-        <svg className="hero-wave" width="100%" height="80" viewBox="0 0 1440 80" fill="none" xmlns="http://www.w3.org/2000/svg">
-          <path d="M0,40 Q360,0 720,40 Q1080,80 1440,40" stroke="#bbb" strokeWidth="2" fill="none"/>
-        </svg>
-      </section>
-      <footer>
-        <div className="social-buttons">
-          <a href="https://facebook.com" target="_blank" rel="noopener noreferrer" className="social-button">
-            <i className="fab fa-facebook"></i>
-          </a>
-          <a href="https://instagram.com" target="_blank" rel="noopener noreferrer" className="social-button">
-            <i className="fab fa-instagram"></i>
-          </a>
-          <a href="mailto:contacto@noesis.com" className="social-button">
-            <i className="fas fa-envelope"></i>
-          </a>
-          <a href="https://linkedin.com" target="_blank" rel="noopener noreferrer" className="social-button">
-            <i className="fab fa-linkedin"></i>
-          </a>
-        </div>
-        <small>© {new Date().getFullYear()} NOESIS. Todos los derechos reservados.</small>
-      </footer>
+      <Contacto/>
+      <Footer/>
     </div>
   );
 }
