+++ src/data/proyectos.ts
@@ -1,0 +1,24 @@
+export const proyectosIniciales = [
+  {
+    nombre: 'Plataforma de gestión educativa',
+    descripcion: 'Sistema web para administración de colegios y universidades.',
+    imagen: 'https://albert-z00z.github.io/Portfolio1/images/NumSecret.png',
+    tecnologias: ['React', 'Node.js', 'MongoDB'],
+    url: 'https://albert-z00z.github.io/juego-secreto/'
+  },
+  {
+    nombre: 'App de reservas para restaurantes',
+    descripcion: 'Aplicación móvil para gestionar reservas y pedidos en restaurantes.',
+    imagen: 'https://via.placeholder.com/320x180/cccccc/888888?text=Proyecto+2',
+    tecnologias: ['React Native', 'Firebase'],
+    url: 'https://ejemplo2.com'
+  },
+  {
+    nombre: 'Integración de sistemas ERP',
+    descripcion: 'Solución personalizada para integrar procesos empresariales.',
+    imagen: 'https://via.placeholder.com/320x180/cccccc/888888?text=Proyecto+3',
+    tecnologias: ['Python', 'Django', 'PostgreSQL'],
+    url: 'https://ejemplo3.com'
+  },
+  // Para agregar más proyectos, seguir el formato de los anteriores
+];
