+++ src/data/testimonios.ts
@@ -1,0 +1,22 @@
+export const testimoniosIniciales = [
+  {
+    nombre: 'Juan Pérez',
+    texto: 'El equipo de NOESIS superó nuestras expectativas. Su atención al detalle y profesionalismo fue excepcional.'
+  },
+  {
+    nombre: 'María Gómez',
+    texto: 'Gracias a NOESIS, nuestra empresa ahora cuenta con una plataforma digital robusta y moderna.'
+  },
+  {
+    nombre: 'Carlos Ruiz',
+    texto: 'Excelente servicio y acompañamiento en todo el proceso de desarrollo. ¡Muy recomendados!'
+  },
+  {
+    nombre: 'Ana Torres',
+    texto: 'NOESIS nos ayudó a digitalizar procesos clave y mejorar la eficiencia de nuestro equipo.'
+  },
+  {
+    nombre: 'Luis Fernández',
+    texto: 'Gran calidad humana y técnica. Siempre atentos a nuestras necesidades.'
+  }
+];
